+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Current (in progress)
 
+- Nothing yet
+
+## 6.1.2 (2023-03-28)
+
+- Resources schema can now have an URL field. [#2825](https://github.com/opendatateam/udata/pull/2825)
 - Fix URLField validation error message [#2831](https://github.com/opendatateam/udata/pull/2831)
 - Dates renaming to provide more relevant computed dates between internal and harvested dates [#2815](https://github.com/opendatateam/udata/pull/2815):
     - Removed `published` attribute from Resource model
@@ -15,6 +20,8 @@
     - Upgrade WTForms 2.2.1 -> 3.0.1
     - Upgrade WTForms-json 0.3.3 -> 0.3.5
     - New security email template for existing users
+- Fix dcat harvesting on dcat:Dataset with blank nodes [#2834](https://github.com/opendatateam/udata/pull/2834)
+- Add dataset archived field in resource catalog [#2833](https://github.com/opendatateam/udata/pull/2833)
 
 ## 6.1.1 (2023-03-17)